// src/generated/client/worldmonitor/conflict/v1/service_client.ts
export interface IranEvent {
  id: string;
  title: string;
  category: string;
  sourceUrl: string;
  latitude: number;
  longitude: number;
  locationName: string;
  timestamp: string;
  severity: string;
}